Провела рефакторинг, разбила на модули и создала класс
