+++ index.html
@@ -67,6 +67,6 @@
 </div>
 <div class="catch_screen"></div>
 
-<script src="main.js"></script>
+<script src="js/main.js" type="module"></script>
 </body>
 </html>
