+++ js/random.js
@@ -1,0 +1,3 @@
+const random = (num) => Math.ceil(Math.random() * num);
+
+export default random;
